zna-streamlit-cloud/zna-repo/requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
reportlab>=4.2.0
tenacity>=8.3.0
pandas>=2.0.0
